# api/openapi/techstack-v1.yaml
openapi: 3.1.0
info:
  title: kombify-TechStack API
  description: |
    REST API for kombify-TechStack -- the hybrid cloud control plane for homelabs.

    kombify-TechStack manages infrastructure through a spec-driven architecture:
    validate YAML specs via CUE, generate OpenTofu configurations, and manage
    nodes through a gRPC agent protocol.

    ## Authentication

    Authenticated endpoints accept an OIDC/JWKS bearer token or an API key
    where documented. Self-hosted TechStack defaults to Pocket ID as identity
    provider. SaaS routes enter through the edge with Auth0 as the platform
    identity provider. PocketBase session exchange remains legacy migration
    compatibility only.

    ## Response Format

    **Success:** `{ "data": <object>, "meta": { "total": int, "page": int, "per_page": int } }`

    **Error:** `{ "error": { "code": "<ERROR_CODE>", "message": "...", "details": <any> } }`

    ## Pagination

    List endpoints accept `page` (default: 1) and `per_page` (default: 20, max: 100) query parameters.
  version: 1.0.0
  contact:
    name: kombify
    url: https://kombify.io
  license:
    name: Apache-2.0 AND GPL-3.0
    url: https://github.com/kombifyio/TechStack/blob/main/LICENSE

# ============================================================
# SERVERS
# ============================================================
# All paths in this document are authored as /api/v1/*. The Gateway projects
...
servers:
  - url: https://api.kombify.io/v1/techstack
...

...
security:
  - bearerAuth: []
paths:
...
  /api/v1/monthly-runtimes/offerings:
    get:
      operationId: listMonthlyRuntimeOfferings
      tags: [Monthly Runtime]
      summary: List monthly runtime offerings
      description: Returns public, provider-free subscription runtime offerings.
      responses:
        "200":
          description: Monthly runtime offerings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MonthlyRuntimeOffering"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/monthly-runtimes/{lease_id}:
    get:
      operationId: getMonthlyRuntimeStatus
      tags: [Monthly Runtime]
      summary: Get monthly runtime status
      parameters:
        - $ref: "#/components/parameters/monthlyRuntimeLeaseId"
      responses:
        "200":
          description: Runtime status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Enrollment is not complete
        "503":
          description: Simulate runtime client is unavailable
  /api/v1/monthly-runtimes/{lease_id}/start:
    post:
      operationId: startMonthlyRuntime
      tags: [Monthly Runtime]
      summary: Start a stopped managed server of the same generation
      description: >
        Admits one generation-bound provider-control reconcile toward running.
        The provider power-on runs once under an at-most-once claim; a repeated
        request while it converges replays the same operation. Requires the
        managed runtime entitlement before any provider call.
      parameters:
        - $ref: "#/components/parameters/monthlyRuntimeLeaseId"
      responses:
        "200":
          description: The power reconcile converged; power.status is succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "202":
          description: The power reconcile is durable and still converging.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "403":
          description: Entitlement denial with error_code, reason_code, missing features and user_guidance.
        "409":
          description: >
            A different power transition is still converging
            (managed_power_transition_in_progress) or the runtime does not
            admit an in-place power change (managed_power_not_admitted).
  /api/v1/monthly-runtimes/{lease_id}/stop:
    post:
      operationId: stopMonthlyRuntime
      tags: [Monthly Runtime]
      summary: Stop a managed server without destroying it
      description: >
        Admits one generation-bound provider-control reconcile toward stopped.
        centron uses its graceful guest shutdown action; IONOS shuts the guest
        down over the execution channel because the IONOS provider stop
        releases the bound DHCP IPv4. The server keeps its disk, address,
        lease, capacity reservation and monthly billing.
      parameters:
        - $ref: "#/components/parameters/monthlyRuntimeLeaseId"
      responses:
        "200":
          description: The power reconcile converged; power.status is succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "202":
          description: The power reconcile is durable and still converging.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "403":
          description: Entitlement denial with error_code, reason_code, missing features and user_guidance.
        "409":
          description: A power transition is converging or the runtime does not admit a pause.
  /api/v1/monthly-runtimes/{lease_id}/enable-ssh:
    post:
      operationId: enableMonthlyRuntimeSSH
      tags: [Monthly Runtime]
      summary: Enable owner SSH access for this generation
      description: >
        Re-enables the generation-bound owner SSH grant over the pinned
        execution channel and reinstalls the keys the owner authorized for this
        generation. New keys are authorized only through the re-authenticated
        server access route.
      parameters:
        - $ref: "#/components/parameters/monthlyRuntimeLeaseId"
      responses:
        "200":
          description: Node-verified grant change; see ssh_access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "403":
          description: Entitlement or demo-account denial.
        "409":
          description: The runtime is stopped (managed_runtime_stopped).
        "503":
          description: The execution channel could not reach the node; no node change was made.
  /api/v1/monthly-runtimes/{lease_id}/disable-ssh:
    post:
      operationId: disableMonthlyRuntimeSSH
      tags: [Monthly Runtime]
      summary: Disable owner SSH access for this generation
      description: >
        Removes every owner-authorized key from the node while keeping
        Kombify's execution-channel key, and records the disabled grant so new
        owner keys are refused until SSH is enabled again.
      parameters:
        - $ref: "#/components/parameters/monthlyRuntimeLeaseId"
      responses:
        "200":
          description: Node-verified grant change; see ssh_access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "403":
          description: Entitlement denial.
        "409":
          description: The runtime is stopped (managed_runtime_stopped).
        "503":
          description: The execution channel could not reach the node; no node change was made.
  /api/v1/monthly-runtimes/{lease_id}/decommission:
    post:
      operationId: decommissionMonthlyRuntime
      tags: [Monthly Runtime]
      summary: Decommission a monthly runtime lease
      parameters:
        - $ref: "#/components/parameters/monthlyRuntimeLeaseId"
        - name: force
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: >
            Request exact generation-bound provider reconciliation when the
            runtime is unreachable. Force is accepted only when the control
            plane can prove restart-safe, HA-fenced durable custody before it
            cancels the lease. Also accepted as a {"force": true} JSON body. An
            empty request defaults to false.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                force:
                  type: boolean
                  default: false
                expected_resource_generation_digest:
                  type: string
                  pattern: "^[0-9a-f]{64}$"
                  description: >
                    Optional optimistic concurrency claim. Provider teardown is
                    refused with 409 unless this digest still identifies the
                    current lease resource generation.
      responses:
        "200":
          description: Confirmed synchronous runtime action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "202":
          description: >
            Durable exact-generation provider reconciliation was accepted but
            provider teardown is not yet proven. observed_state and lease_state
            are reconciliation_pending; this is not terminal success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "400":
          description: >
            Invalid, oversized, ambiguous, or schema-incompatible JSON body,
            or an invalid expected resource-generation digest.
        "409":
          description: >
            Decommission blocked because the runtime is unreachable or because
            the expected resource generation has been superseded. Unreachable
            errors carry error_code=decommission_blocked_unreachable with a
            structured user_guidance offering force=true.
        "503":
          description: >
            The Simulate runtime client is unavailable, or restart-safe and
            HA-fenced durable reconciliation custody is not available. In the
            latter case no lease cancellation, generation claim, or pending
            metadata mutation is performed.
  /api/v1/monthly-runtimes/{lease_id}/resolve-custody:
    post:
      operationId: resolveMonthlyRuntimeCustody
      tags: [Monthly Runtime]
      summary: Archive a legacy or unbound custody record after provider cleanup
      description: >
        Removes a non-executable legacy or unbound lease from the active
        dashboard only after the owner confirms that its provider resource was
        already removed. This operation never contacts or deletes a provider
        resource and is rejected for Techstack provider-controlled generations.
      parameters:
        - $ref: "#/components/parameters/monthlyRuntimeLeaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [provider_cleanup_confirmed]
              properties:
                provider_cleanup_confirmed:
                  type: boolean
                  enum: [true]
      responses:
        "200":
          description: Custody record archived with owner-confirmation evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "400":
          description: Invalid or schema-incompatible confirmation body
        "409":
          description: >
            Provider cleanup was not confirmed, or the lease remains under
            Techstack provider control and must use the decommission lifecycle.
...
  /api/v1/monthly-runtimes/{lease_id}/ssh:
    get:
      operationId: getMonthlyRuntimeSSHInfo
      tags: [Monthly Runtime]
      summary: Get SSH connection information for a monthly runtime lease
      parameters:
        - $ref: "#/components/parameters/monthlyRuntimeLeaseId"
      responses:
        "200":
          description: SSH info result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRuntimeEnvelope"
        "503":
          description: Simulate runtime client is unavailable
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC/JWKS bearer token. Self-hosted defaults to Pocket ID; SaaS uses Auth0 through the edge. PocketBase JWT is legacy migration compatibility only.
    runtimeAgentAuth:
      type: http
      scheme: bearer
      bearerFormat: TechStack-Runtime-Agent
      description: Runtime-agent credential bound to the exact tenant, placement, server, and worker identity. Signed, opaque, and explicit legacy compatibility tokens are validated fail-closed by the worker runtime.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service accounts and automation clients (SaaS mode via the public edge)
    edgeTrustHeader:
      type: apiKey
      in: header
      name: X-Edge-Internal-Trust
      description: Legacy trust header injected by the public edge during the migration cutover (SaaS mode only)
...